+++ docs/changes.rst
@@ -23,6 +23,7 @@ Weblate 4.6
 
 Released on April 19th 2021.
 
+* Added support for ``object-pascal-format`` used in gettext PO.
 * The auto_translate management command has now a parameter for specifying translation mode.
 * Added support for :ref:`txt`.
 * Added trends and metrics for all objects.